+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.4.2] - 2019-02-21
+### Added
+- Config file with current and default settings
+- Settings are now remembered 
+
+### Fixed
+- No more error message while closing game
+
 ## [0.4.1] - 2019-02-19
 ### Added
-- execution (.exe) file
+- Execution (.exe) file
 - file with fonts
 
 ### Fixed